refactor(models): type trade status as a shared TradeStatus literal
