+++ src/polymarket/models/__init__.py
@@ -36,6 +36,7 @@ from polymarket.models.clob import (
     SignedOrder,
     TickSize,
     TotalUserEarning,
+    TradeStatus,
     TransactionOutcome,
     UserEarning,
     UserRewardsConfig,
@@ -204,6 +205,7 @@ __all__ = [
     "BuilderTrade",
     "CancelOrdersResponse",
     "ClobTrade",
+    "TradeStatus",
     "CurrentReward",
     "CurrentRewardConfig",
     "EarningBreakdown",
